Implement APIv2 (/metadata)
This API is more efficient and UGS will request every change in view
when using v2, so it's a much better user experience (instead of just
showing the badges that appear after you started the application).

This still supports /build, this is mostly just because a running client
won't re-check `/latest` to find the API version until it's restarted,
but it's also gives us back compat.