Switch to using YAML for the configuration
Originally I used JSON to not add additional dependencies (since it's
used by the web API), but JSON is just kind of a tedious format for
configuration files. It will now try to load the config passed via
`--config` as YAML, and if it isn't passed, it will try loading
`config.yaml`, and if *that* doesn't exist, it will try loading
`config.json` so that old installations will continue to work.