Dockerfile: Don't copy all the busybox libs
This copies hundreds of 1mb files, which isn't a problem when you use it
as a base image because they're either symlinks or hardlinks to the same
binary, but `COPY /bin*` copies them as regular files, bloating the
image.

This reduces the size back down. The other tools were only use for
debugging. In the future, we should investigate using e.g. musl and
alpine, so we can get package management and still retain a small size.